파일 삭제 및 코드 수정

+++ Spring/src/main/java/com/heaildairy/www/recommend/movie/moviedto/EmotionGenreMapEntityDTO.java
@@ -11,5 +11,5 @@ public class EmotionGenreMapEntityDTO {
     private String emotionType;
     private Integer genreCode;
     private String genreName;
-    private Integer genreWeight;
+    private Double genreWeight;
 }
